requirements.txt
tqdm
streamlit
llama-index-agent-openai
llama-index-embeddings-openai
llama-index-llms-openai
llama-index==0.10.12
python-dotenv
llama-index-agent-openai-legacy
llama-index-postprocessor-cohere-rerank
llama-index-llms-huggingface
transformers
sentence_transformers
pinecone-client
numpy
llama-index-embeddings-nomic